# [REVIEW]Optimize Embedding Creation by 4.8x