--- a/cuBERT_topic_modelling/embedding_extraction.py
+++ b/cuBERT_topic_modelling/embedding_extraction.py
@@ -1,14 +1,14 @@
-from cudf.core.subword_tokenizer import SubwordTokenizer
+from cudf.core.subword_tokenizer import SubwordTokenizer, _cast_to_appropriate_type
 import torch
-from transformers import AutoModel
 from torch.utils.data import TensorDataset, DataLoader
+
+import time
 
 # Vocabulary is included in the root directory of this repo
 # however, below is the command to modify / update it -->
 # from cudf.utils.hash_vocab_utils import hash_vocab
 # hash_vocab('vocab.txt', 'voc_hash.txt')
 
-device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
 # Mean Pooling - Take attention mask into account for correct averaging
 def mean_pooling(model_output, attention_mask):
@@ -34,7 +34,33 @@ def mean_pooling(model_output, attention_mask):
     return sum_embeddings / sum_mask
 
 
-def create_embeddings(sentences):
+def tokenize_strings(sentences, tokenizer):
+    max_length = 128
+
+    # Tokenize cudf Series
+    token_o = tokenizer(
+        sentences,
+        max_length=max_length,
+        max_num_rows=len(sentences),
+        padding="max_length",
+        return_tensors="cp",
+        truncation=True,
+        add_special_tokens=True,
+    )
+
+    clip_len = max_length - int((token_o["input_ids"][:, ::-1] != 0).argmax(1).min())
+    token_o["input_ids"] = _cast_to_appropriate_type(
+        token_o["input_ids"][:, :clip_len], "pt"
+    )
+    token_o["attention_mask"] = _cast_to_appropriate_type(
+        token_o["attention_mask"][:, :clip_len], "pt"
+    )
+
+    del token_o["metadata"]
+    return token_o
+
+
+def create_embeddings(sentences, embedding_model, vocab_file="vocab/voc_hash.txt"):
     """Creates the sentence embeddings using SentenceTransformer
 
     Args:
@@ -45,44 +71,26 @@ def create_embeddings(sentences):
         embeddings for the strings passed
     """
 
-    cudf_tokenizer = SubwordTokenizer("vocab/voc_hash.txt", do_lower_case=True)
-
-    # Tokenize sentences
-    encoded_input_cudf = cudf_tokenizer(
-        sentences,
-        max_length=128,
-        max_num_rows=len(sentences),
-        padding="max_length",
-        return_tensors="pt",
-        truncation=True,
-    )
-
-    # Load AutoModel from huggingface model repository
-    model_gpu = AutoModel.from_pretrained("sentence-transformers/all-MiniLM-L6-v2").to(
-        device
-    )
-
-    # Delete the key and associated values we do not need
-    del encoded_input_cudf["metadata"]
-
+    cudf_tokenizer = SubwordTokenizer(vocab_file, do_lower_case=True)
     batch_size = 64
-
-    dataset = TensorDataset(
-        encoded_input_cudf["input_ids"], encoded_input_cudf["attention_mask"]
-    )
-
-    dataloader = DataLoader(dataset, batch_size=batch_size)
-
-    del encoded_input_cudf["input_ids"], encoded_input_cudf["attention_mask"]
-
     pooling_output_ls = []
+    model_st = time.time()
     with torch.no_grad():
-        for b_input_ids, b_attention_mask in dataloader:
-            model_obj = model_gpu(
+        for s_ind in range(0, len(sentences), batch_size):
+            e_ind = min(s_ind + batch_size, len(sentences))
+            b_s = sentences[s_ind:e_ind]
+
+            tokenized_d = tokenize_strings(b_s, cudf_tokenizer)
+            b_input_ids = tokenized_d["input_ids"]
+            b_attention_mask = tokenized_d["attention_mask"]
+
+            model_obj = embedding_model(
                 **{"input_ids": b_input_ids, "attention_mask": b_attention_mask}
             )
             pooling_output_ls.append(mean_pooling(model_obj, b_attention_mask))
 
     pooling_output = torch.cat(pooling_output_ls)
+    model_et = time.time()
+    print(f"DL time = {model_et-model_st}")
 
     return pooling_output